Implement preliminary support for RFC3894 Copy
Implement the :copy argument for fileinto and redirect in the parser as
specified in RFC3894 "Copying Without Side Effects".

Note this only adds it to the parser, but does not expose it in the UI.
It works similarly to the existing :create argument, which is also not
supported in the UI, but preserved on editing.

The implementation is not perfect, as the order of arguments to fileinto
is hardcoded. If both arguments should be used, the statement needs to
be written as 'fileinto :copy :create "mailbox";'.

See: #61