Fix hang on fast responding servers (#638)
The message queue gets of sync in case multiple requestes where queued and parsing the first request failed. E.g. is was inclompleted because the server responded fast or in fragments.

In this secenario the request queue was not reset and thus the request got removed from the message queue despite having failed to parse the response.

This makes the server and client go out of sync which results in a hang.